pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-cot-mutator"
version = "0.1.0"
description = "Chain-of-Thought mutation experiments for evaluating model robustness"
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0.2",
    "pytest>=8.3.4",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
huggingface = ["transformers>=4.30.0", "torch>=2.0.0"]

[project.scripts]
chat-cot-runner = "tools.runner:main"

[tool.setuptools.packages.find]
include = ["clients*", "core*", "eval*", "mutations*", "datasets*", "tools*", "app_components*"]
